AI-1172 refactor: move global_search tool to its module representing search tools
